requirements.txt
discord
PyQt6
colorama
pytest
pillow